requirements.txt
unstructured-ingest
unstructured-ingest[remote]
python-dotenv
graphrag_sdk
networkx
matplotlib
google-generativeai